2.4.3 b
Changed "aid" label to "pre" label to prevent future confusion. In KERI events the element is the autonomic identifier prefix only not the full autonomic identifier. Later in other parts of KERI when will start using full aids it will be confusing that the event field label for merely the prefix is aid not something more evocative of prefix like "pre"